change helper, c3, c5
change execution time type
change input for challenge 3
change decimal to x/y for challenge 5

+++ src/helper/ExecutionTimeHelper.java
@@ -18,6 +18,6 @@ public class ExecutionTimeHelper {
      * @param startTime long
      */
     public static void printExecutionTime(long startTime) {
-        System.out.format("%nExecution time: %d ms", (System.nanoTime() - startTime) / 1000000);
+        System.out.format("%nExecution time: %.2f ms", (double)(System.nanoTime() - startTime) / 1000000);
     }
 }
